+++ src/app/budget/components/line-menu/line-menu.service.ts
@@ -8,10 +8,12 @@ export class LineMenuService {
   // Observable string sources
   private addExpenseSource = new Subject<any>();
   private addCashSource = new Subject<any>();
+  private moreSource = new Subject<any>();
 
   // Observable string streams
   addExpense$ = this.addExpenseSource.asObservable();
   addCash$ = this.addCashSource.asObservable();
+  more$ = this.moreSource.asObservable();
 
   emitAddExpense(budgetLineId: string) {
     this.addExpenseSource.next(budgetLineId);
@@ -19,5 +21,9 @@ export class LineMenuService {
 
   emitAddCash(budgetLineId: string) {
     this.addCashSource.next(budgetLineId);
+  }
+
+  emitMore(budgetLineId: string){
+    this.moreSource.next(budgetLineId);
   }
 }