+++ src/app/budget/components/lines-list/lines-list.component.ts
@@ -9,6 +9,7 @@ import {
 } from '@angular/core';
 import { BudgetLine } from '../../models/budget-line.model';
 import { Subject, Subscription } from 'rxjs';
+import { SortingLinesTypes } from '../lines-sort/lines-sort.component';
 
 @Component({
   selector: 'app-lines-list',
@@ -19,64 +20,28 @@ import { Subject, Subscription } from 'rxjs';
 export class LinesListComponent implements OnInit, OnDestroy {
   @Input() lines: BudgetLine[] = [];
 
-  @Input() sortSubject: Subject<string>;
-
-  @Input() sortBy: string;
-
   @Input() isFastExpenseMode: boolean;
+
+  @Input() sortBy: SortingLinesTypes;
 
   @Output() addExpenseEmitter = new EventEmitter<any>();
 
   @Output() selectLine = new EventEmitter<string>();
 
-  @Output() sortLines = new EventEmitter<string>();
-
-  sortSubcription: Subscription;
+  @Output() sortLines = new EventEmitter<SortingLinesTypes>();
 
   constructor() {}
 
   ngOnInit() {
-    this.sortSubcription = this.sortSubject.subscribe((sortBy: string) => {
-      this.sortBy = sortBy;
-    });
   }
 
   private onSelectLine(budgetLineId: string) {
     this.selectLine.emit(budgetLineId);
   }
 
-  onSortLines(sortBy: string) {
+  onSortLines(sortBy: SortingLinesTypes) {
     this.sortLines.emit(sortBy);
   }
 
-  sortLinesBy(sortBy: string): BudgetLine[] {
-    switch (sortBy) {
-      case 'alphaUp':
-        return this.lines.sort((a: BudgetLine, b: BudgetLine) => {
-          return a.name.toLowerCase().localeCompare(b.name.toLowerCase());
-        });
-      case 'alphaDown':
-        return this.lines.sort((a: BudgetLine, b: BudgetLine) => {
-          return b.name.toLowerCase().localeCompare(a.name.toLowerCase());
-        });
-      case 'amountUp':
-        return this.lines.sort((a: BudgetLine, b: BudgetLine) => {
-          return a.cashLeft - b.cashLeft;
-        });
-      case 'amountDown':
-        return this.lines.sort((a: BudgetLine, b: BudgetLine) => {
-          return b.cashLeft - a.cashLeft;
-        });
-      default:
-        return this.lines.sort((a: BudgetLine, b: BudgetLine) => {
-          return a.name.toLowerCase().localeCompare(b.name.toLowerCase());
-        });
-    }
-  }
-
-  ngOnDestroy() {
-    if (this.sortSubcription) {
-      this.sortSubcription.unsubscribe();
-    }
-  }
+  ngOnDestroy() {}
 }