+++ src/app/budget/components/lines-sort/lines-sort.component.ts
@@ -1,10 +1,17 @@
-import { Component, OnInit, EventEmitter, Output } from '@angular/core';
+import { Component, OnInit, EventEmitter, Output, Input } from '@angular/core';
 import {
   faSortAmountDown,
   faSortAmountUp,
   faSortAlphaDown,
   faSortAlphaUp
 } from '@fortawesome/free-solid-svg-icons';
+
+export enum SortingLinesTypes {
+  AMOUNT_DOWN = '[Sort Lines] Amount Down',
+  AMOUNT_UP = '[Sort Lines] Amount Up',
+  ALPHA_DOWN = '[Sort Lines] Alpha Down',
+  ALPHA_UP = '[Sort Lines] Alpha Up'
+}
 
 @Component({
   selector: 'app-lines-sort',
@@ -17,18 +24,40 @@ export class LinesSortComponent implements OnInit {
   faSortAlphaUp = faSortAlphaUp;
   faSortAlphaDown = faSortAlphaDown;
 
+  @Input() sortBy: SortingLinesTypes;
 
   selectedValue: string;
 
-  @Output() sortLines = new EventEmitter<string>();
+  @Output() sortLines = new EventEmitter<SortingLinesTypes>();
 
   constructor() {}
 
   ngOnInit() {
+    this.selectedValue = this.sortBy;
+  }
+
+  isChecked(value: string) {
+    if (value && value === this.sortBy) {
+      console.log('isChecked is true for value: ', value);
+    }
   }
 
   onChange(newValue: string) {
     this.selectedValue = newValue;
-    this.sortLines.emit(this.selectedValue);
+    switch (this.selectedValue) {
+      case SortingLinesTypes.ALPHA_DOWN:
+        this.sortBy = SortingLinesTypes.ALPHA_DOWN;
+        break;
+      case SortingLinesTypes.ALPHA_UP:
+        this.sortBy = SortingLinesTypes.ALPHA_UP;
+        break;
+      case SortingLinesTypes.AMOUNT_DOWN:
+        this.sortBy = SortingLinesTypes.AMOUNT_DOWN;
+        break;
+      case SortingLinesTypes.AMOUNT_UP:
+        this.sortBy = SortingLinesTypes.AMOUNT_UP;
+        break;
+    }
+    this.sortLines.emit(this.sortBy);
   }
 }