+++ src/app/budget/models/budget-line.model.ts
@@ -3,4 +3,5 @@ export interface BudgetLine {
     cashLeft: number;
     cashToSpend: number;
     name: string;
+    description?: string;
 }