+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.37.0
+* Improved sorting budget lines
+* First sketch of line details
+
 ## 0.36.0
 * Sorting budget lines